requirements.txt
black==22.8.0
clearml==1.7.0
clearml-agent==1.3.0
flake8==5.0.4
isort==5.10.1
numpy==1.23.3
opencv-python==4.6.0.66
pandas==1.5.0
pathlib2==2.3.7.post1
pep8-naming==0.13.2
Pillow==9.2.0
pydantic==1.10.2
pydantic-yaml==0.8.0
pytorch-lightning==1.7.7
scikit-learn==1.1.2
scipy==1.9.1
tensorboard==2.10.0
torch==1.12.1+cu113
torchmetrics==0.9.3
torchvision==0.13.1+cu113
tqdm==4.64.1
